+++ src/.internal/charts/xy/XYChart.ts
@@ -537,8 +537,8 @@ export class XYChart extends SerialChart {
 						position = wheelZoomPositionX;
 					}
 
-					let newStart = start - wheelStep * (end - start) * shiftX * position;
-					let newEnd = end + wheelStep * (end - start) * shiftX * (1 - position);
+					let newStart =  Math.min(1, Math.max(0, start - wheelStep * (end - start) * shiftX * position));
+					let newEnd = Math.max(0, Math.min(1, end + wheelStep * (end - start) * shiftX * (1 - position)));
 					if (1 / (newEnd - newStart) < axis.getPrivate("maxZoomFactor", Infinity) / axis.get("minZoomCount", 1)) {
 						this._handleWheelAnimation(axis.zoom(newStart, newEnd));
 					}
@@ -558,8 +558,8 @@ export class XYChart extends SerialChart {
 						position = wheelZoomPositionX;
 					}
 
-					let newStart = start - wheelStep * (end - start) * shiftY * position;
-					let newEnd = end + wheelStep * (end - start) * shiftY * (1 - position);
+					let newStart = Math.min(1, Math.max(0, start - wheelStep * (end - start) * shiftY * position));
+					let newEnd = Math.max(0, Math.min(1, end + wheelStep * (end - start) * shiftY * (1 - position)));
 
 					if (1 / (newEnd - newStart) < axis.getPrivate("maxZoomFactor", Infinity) / axis.get("minZoomCount", 1)) {
 						this._handleWheelAnimation(axis.zoom(newStart, newEnd));
@@ -581,8 +581,8 @@ export class XYChart extends SerialChart {
 						position = wheelZoomPositionY;
 					}
 
-					let newStart = start - wheelStep * (end - start) * shiftX * position;
-					let newEnd = end + wheelStep * (end - start) * shiftX * (1 - position);
+					let newStart = Math.min(1, Math.max(0, start - wheelStep * (end - start) * shiftX * position));
+					let newEnd = Math.max(0, Math.min(1, end + wheelStep * (end - start) * shiftX * (1 - position)));
 
 					if (1 / (newEnd - newStart) < axis.getPrivate("maxZoomFactor", Infinity) / axis.get("minZoomCount", 1)) {
 						this._handleWheelAnimation(axis.zoom(newStart, newEnd));
@@ -603,8 +603,8 @@ export class XYChart extends SerialChart {
 						position = wheelZoomPositionY;
 					}
 
-					let newStart = start - wheelStep * (end - start) * shiftY * position;
-					let newEnd = end + wheelStep * (end - start) * shiftY * (1 - position);
+					let newStart = Math.min(1, Math.max(0, start - wheelStep * (end - start) * shiftY * position));
+					let newEnd = Math.max(0, Math.min(1, end + wheelStep * (end - start) * shiftY * (1 - position)));
 
 					if (1 / (newEnd - newStart) < axis.getPrivate("maxZoomFactor", Infinity) / axis.get("minZoomCount", 1)) {
 						this._handleWheelAnimation(axis.zoom(newStart, newEnd));