+++ src/.internal/charts/xy/axes/AxisRenderer.ts
@@ -84,6 +84,14 @@ export interface IAxisRendererSettings extends IGraphicsSettings {
 	 */
 	pan?: "none" | "zoom"
 
+
+	/**
+	 * Sensitivity of panning. The higher the number, the more sensitive it is.
+	 *
+	 * @default 1
+	 */
+	panSensitivity?: number;
+
 }
 
 export interface IAxisRendererPrivate extends IGraphicsPrivate {
@@ -365,7 +373,7 @@ export abstract class AxisRenderer extends Graphics {
 
 			const downStart = this._downStart!;
 			const downEnd = this._downEnd!;
-			const extra = this._getPan(point, downPoint) * Math.min(1, (downEnd - downStart)) / 2;
+			const extra = this._getPan(point, downPoint) * Math.min(1, (downEnd - downStart)) / 2 * this.get("panSensitivity", 1);
 			this.axis.zoom(downStart - extra, downEnd + extra, 0);
 		}
 	}